requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2-binary==2.8.5
requests==2.23.0
six==1.14.0
urllib3==1.25.9
vk-api==11.8.0
Werkzeug==1.0.1
whitenoise==5.0.1
